Clean up types
