+++ public/app/plugins/datasource/grafana-azure-monitor-datasource/datasource.ts
@@ -75,6 +75,8 @@ export default class Datasource extends DataSourceApi<AzureMonitorQuery, AzureDa
   }
 
   query(options: DataQueryRequest<AzureMonitorQuery>): Observable<DataQueryResponse> {
+    console.log('query options.targets', options.targets);
+
     const byType = new Map<AzureQueryType, DataQueryRequest<AzureMonitorQuery>>();
 
     for (const target of options.targets) {
